+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog for NeoFS Node
 - Neo RPC client failure when the first endpoint is unavailable even if there are more endpoints to try (#2703)
 - Incorrect address mapping of the Alphabet contracts in NNS produced by deployment procedure (#2713)
 - IR compares and logs public keys difference, not hash of keys difference at SN verification check (#2711)
+- Incorrect handling of notary request leading to inability to collect signatures in some cases (#2715)
 
 ### Changed
 - Created files are not group writable (#2589)